+++ core/components/frontbar/docs/changelog.md
@@ -1,3 +1,11 @@
+# Frontbar 1.2.0-beta (June 11, 2019)
+========================================
+- Fixed inconsistent manager session check
+- Dropped JS and JQuery
+- Dropped Semantic UI dependency
+- Switched to CSS-only Bulma for styling
+- Additional code cleanup
+
 # Frontbar 1.1.0-beta (September 11, 2016)
 ========================================